codify 2^64 DER payload bytes limit
theoretically, a DER TLV's length field can contain 127 bytes of usable
length information, i.e., an integer with max value 2^(8*127 + 1) - 1.

using, e.g., BigInt, it is of course feasable to represent integers this
large.  However, I strongly believe that no one will even think to
attempt to decode ASN.1 DER in excess of the currently supported 2^63 =
8 Exabytes in the forseable future anyways.  Therefore, it is more
likely that a length field this large is erroneous and we should (and
are) therefore guarding against this via an exception. For now, TLV's
length is restricted to the platforms standard Int size without the
sign bit as negative length's obviously don't make sense.